Allocate stations across strata

Allocates stations across strata based on: stratum size, historical standard devation of cpue, ex vessel price, and abundance of a preselected subset of species, and the time it takes to complete tows. (Reference a tech memo for how exactly Neyman Allocations are handled?).

Updates of code from Ned: Initial changes are to reformat code into my preferred configuration Added ", believeNRows = FALSE" to odbc calls for version 3.0.1 of R

Added package XLConnect to provide functionality to to read Excel workbooks under version 3.0.1 of R in 64-bit environment

Created 12/18/13

Usage

allocate.effort(
  survey,
  channel = NA,
  schema,
  pwrd,
  points,
  stations.available,
  number.of.tows = 741,
  output.file
)

Arguments

survey

character string. Defaults to "AI" for Aleutian Islands

channel

open connection object. Created from RODBC::odbcConnect()

schema

character string. SQL schema name. Defaults to AIGOA_WORK_DATA

pwrd

character string. SQL password

points

dataframe of potential station locations queried from SQL, called in AIGOASurveyPlanning::goa.planning().

stations.available

integer. Vector of available stations in each stratum.

number.of.tows

integer. Number of total stations.

output.file

character string. Output directory.

Value

Author

Ned Laman ned.laman@noaa.gov