Gulf of Alaska Bottom Trawl Survey: Allocate stations across strata

Allocates stations under the Gulf of Alaska stratified random design, last restratified in 2023. Species information is incorporated from the survey years between 1996-2021. Bethel algorithm is used to conduct the multispecies station allocation.

Usage

goa_allocate_stations(
  n = 550,
  min_n_per_stratum = 4,
  species = c("arrowtooth flounder", "Pacific cod", "walleye pollock", "rex sole",
    "flathead sole", "Pacific halibut", "southern rock sole", "northern rock sole",
    "Pacific ocean perch", "silvergray rockfish", "northern rockfish", "dusky rockfish",
    "BS and RE rockfishes", "Dover sole", "shortspine thornyhead"),
  max_iter = 5000,
  trawl = c("Y", "N", "UNK")[c(1, 3)],
  year = 2025
)

Arguments

n

integer.Total number of stations

min_n_per_stratum

integer. Minimum number of stations to assign a stratum

species

character vector. See example on github readme

max_iter

integer. Maximum number of times for bethel algorithm to run to converge to `n`.

trawl

character vector. Options are "Y" for trawlable, "N" for untrawlable, or "UNK" for unknown trawlablity.

year

integer. The year as an integer is used as a seed for the random number generator when drawing random stations.

Value

A named list with elements: 1) ms_allocation: vector, allocation of `n` stations across strata using information from the `species` vector. 2) drawn_stations: dataframe with `n` row and 15 fields. Notable fields are:

STATIONID

Station ID of the grid the station resides. This format of the station id is XX-YY where XX is the grid location on the W-E axis and YY is the grid location on the N-S axis.

STRATUM

Stratum ID. The 2025 restratification stratum id is formatted as the year the stratification was created (i.e., 2025), followed by "_" and then a three digit code in the format ABC similar to the format used in the historical GOA survey design. The second digit (B) refers to the managment area (1: Shumagin, 2: Chirikof, 3: Kodiak, 4: Yakutat, and 5: Southeastern). The first digit (A) denotes the relative depth of the stratum, 0 being the shallowest stratum and 5 being the deepest stratum. Because each management area has different depth strata, this digit is not comparable among management areas, just within a management area, e.g., 2023_010 is shallower than 2023_110. The third digit C is a replicate counter in case there are multiple strata within a management area x depth bin combination (there are no replicates in the 2023 GOA stratification, so this digit is always zero). This last digit is consistent with the historical GOA strata ids.

AREA_KM2

area of station in km^2

TRAWLABLE_AREA_KM2

trawlable area within station in km^2

PERIMETER_KM

perimeter of station in km

CENTER_LAT, CENTER_LONG

Latitude and longitude of station centroid

Author

Zack Oyafuso zack.oyafuso@noaa.gov